Move server models into mzutils
