+++ src/models/servercountrymodel.cpp
@@ -14,9 +14,9 @@
 #include "feature/feature.h"
 #include "leakdetector.h"
 #include "logger.h"
+#include "models/servercountry.h"
 #include "mozillavpn.h"
 #include "recommendedlocationmodel.h"
-#include "servercountry.h"
 #include "serverdata.h"
 #include "serverlatency.h"
 #include "settingsholder.h"
