+++ tests/qml/CMakeLists.txt
@@ -40,12 +40,6 @@ target_sources(qml_tests PRIVATE
     ${MZ_SOURCE_DIR}/controller.h
     ${MZ_SOURCE_DIR}/models/recommendedlocationmodel.cpp
     ${MZ_SOURCE_DIR}/models/recommendedlocationmodel.h
-    ${MZ_SOURCE_DIR}/models/server.cpp
-    ${MZ_SOURCE_DIR}/models/server.h
-    ${MZ_SOURCE_DIR}/models/servercity.cpp
-    ${MZ_SOURCE_DIR}/models/servercity.h
-    ${MZ_SOURCE_DIR}/models/servercountry.cpp
-    ${MZ_SOURCE_DIR}/models/servercountry.h
     ${MZ_SOURCE_DIR}/models/servercountrymodel.cpp
     ${MZ_SOURCE_DIR}/models/servercountrymodel.h
     ${MZ_SOURCE_DIR}/models/serverdata.cpp
